feat: product controller

+++ src/SmallRetail.Services/ProductService.cs
@@ -27,18 +27,35 @@ namespace SmallRetail.Services
 
         public void Create(Product product)
         {
+            product.DateCreated = DateTime.UtcNow;
+            product.DateUpdated = DateTime.UtcNow;
             _db.Add(product);
             _db.SaveChanges();
         }
 
         public void Update(Product product)
         {
-            _db.Update(product);
+            var existingProduct = _db.Products.Find(product.Id);
+            if (existingProduct == null)
+            {
+                throw new ArgumentException("Product doesn't exists");
+            }
+
+            existingProduct.Barcode = product.Barcode;
+            existingProduct.Name = product.Name;
+            existingProduct.Price = product.Price;
+            existingProduct.DateUpdated = DateTime.UtcNow;
+            _db.Update(existingProduct);
             _db.SaveChanges();
         }
 
-        public void Delete(Product product)
+        public void Delete(Guid id)
         {
+            var product = _db.Products.Find(id);
+            if (product == null)
+            {
+                throw new ArgumentException("Product doesn't exists");
+            }
             _db.Remove(product);
             _db.SaveChanges();
         }
